Move node group balancing to processor
The goal is to allow customization of this logic
for different use-case and cloudproviders.
