+++ cluster-autoscaler/processors/nodegroupset/compare_nodegroups.go
